meson: add '-Wall' and '-Werror'
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1238690
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>

+++ librobinhood/src/plugins/s3/s3_internals.h
@@ -17,7 +17,7 @@
 #include <robinhood/action.h>
 
 struct item_data {
-    int64_t length;
+    size_t length;
     int64_t current_id;
     char **list;
 };
